# docs/api_doc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Library Management Service API",
    "description": "Library Management Service API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/theloai": {
      "get": {
        "tags": [
          "The loai"
        ],
        "description": "Get danh sach the loai",
        "responses": {
          "200": {
            "description": "Get danh sach the loai successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TheLoaiPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "description": "Create new the loai",
        "tags": [
          "The loai"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TheLoaiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create the loai successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TheLoaiPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/theloai/{maTheLoai}": {
      "get": {
        "tags": [
          "The loai"
        ],
        "description": "Get the loai by id",
        "parameters": [
          {
            "in": "path",
            "name": "maTheLoai",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get the loai successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TheLoaiPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "description": "Update the loai successful",
        "tags": [
          "The loai"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maTheLoai",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TheLoaiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update the loai successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TheLoaiPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "The loai"
        ],
        "description": "Delete the loai by id",
        "parameters": [
          {
            "in": "path",
            "name": "maTheLoai",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete the loai successful"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/tacgia": {
      "get": {
        "summary": "Get danh sach tac gia",
        "tags": [
          "Tac gia"
        ],
        "responses": {
          "200": {
            "description": "Get danh sach tac gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TacGiaDto"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create tac gia",
        "tags": [
          "Tac gia"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TacGiaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Get danh sach tac gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TacGiaPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/tacgia/{maTacGia}": {
      "get": {
        "summary": "Get tac gia by ma tac gia",
        "tags": [
          "Tac gia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maTacGia",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get tac gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TacGiaPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Update tac gia by ma tac gia",
        "tags": [
          "Tac gia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maTacGia",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TacGiaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update tac gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TacGiaPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete tac gia by ma tac gia",
        "tags": [
          "Tac gia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maTacGia",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete tac gia successful"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TheLoaiDto": {
        "type": "object",
        "properties": {
          "tenTheLoai": {
            "type": "string"
          }
        }
      },
      "TheLoaiPresenter": {
        "type": "object",
        "properties": {
          "maTheLoai": {
            "type": "string"
          },
          "tenTheLoai": {
            "type": "string"
          }
        }
      },
      "TacGiaDto": {
        "type": "object",
        "properties": {
          "tenTacGia": {
            "type": "string"
          }
        }
      },
      "TacGiaPresenter": {
        "type": "object",
        "properties": {
          "maTacGia": {
            "type": "string"
          },
          "tenTacGia": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "BadRequestError": {
        "description": "provide bad request"
      },
      "ForbiddenError": {
        "description": "do not have permission to view this resource"
      },
      "NotFoundError": {
        "description": "resource not found"
      },
      "UnauthorizedError": {
        "description": "token is missing"
      },
      "ConflictError": {
        "description": "conflict with state of server, maybe due to database constraint reason,.."
      },
      "InternalServerError": {
        "description": "internal server error"
      }
    }
  }
}